add symbol to four hour forecast

+++ Rakett_App/app/src/main/java/no/uio/ifi/in2000/team6/rakett_app/ui/cards/DateCard.kt
@@ -61,7 +61,7 @@ fun DateCard(
                 verticalAlignment = Alignment.CenterVertically
             ) {
                 Text(text = weatherInfo.date, style = MaterialTheme.typography.titleMedium)
-                WeatherRatingIndicator(weatherInfo.weatherScore)
+                WeatherRatingIndicator(weatherInfo.weatherScore.toInt())
             }
 
             Spacer(modifier = Modifier.height(8.dp))
